Move creation of id3 data to a component
