+++ src/api/create.ts
@@ -1,7 +1,7 @@
-import * as ID3Util from "../ID3Util"
 import { WriteTags } from "../types/Tags"
 import { isFunction } from  "../util"
 import { createBufferFromTags } from "../TagsHelpers"
+import { createId3Data } from "../id3-data"
 
 /**
  * Callback used to return a buffer with the created ID Tag.
@@ -26,18 +26,7 @@ export function create(tags: WriteTags): Buffer
 export function create(tags: WriteTags, callback: CreateCallback): void
 
 export function create(tags: WriteTags, callback?: CreateCallback) {
-    const frames = createBufferFromTags(tags)
-
-    //  Creates ID3 header
-    const header = Buffer.alloc(10)
-    header.fill(0)
-    header.write("ID3", 0)              // File identifier
-    header.writeUInt16BE(0x0300, 3)     // Version 2.3.0  --  03 00
-    header.writeUInt16BE(0x0000, 5)     // Flags 00
-    ID3Util.encodeSize(frames.length).copy(header, 6)
-
-    const id3Data = Buffer.concat([header, frames])
-
+    const id3Data = createId3Data(createBufferFromTags(tags))
     if (isFunction(callback)) {
         return callback(id3Data)
     }
